+++ app/src/main/java/com/example/rezepteapp/model/RecipeModel.java
@@ -2,16 +2,29 @@ package com.example.rezepteapp.model;
 
 import android.content.Context;
 
+import com.example.rezepteapp.daos.IngredientDAOImpl;
+import com.example.rezepteapp.daos.ShoppinglistDAOImpl;
 import com.example.rezepteapp.database.RecipeDbOpenHelper;
+import com.example.rezepteapp.mapper.ToShoppingListMapper;
 
 import java.util.List;
 
 public class RecipeModel {
     private Recipe recipe;
-    private RecipeDbOpenHelper db;
+    private List<ShoppinglistEntry> toDoList;
+    private ShoppinglistDAOImpl sldao;
+    private IngredientDAOImpl idao;
+
+
+    public List<ShoppinglistEntry> getToDoList() {
+        toDoList.clear();
+        ToShoppingListMapper mapper = new ToShoppingListMapper();
+        toDoList.addAll(mapper.mapToList(sldao.getAllShoppinglist(), idao.getAllIngredients()));
+
+        return toDoList;
+    }
 
     public RecipeModel(Context context) {
-        db = new RecipeDbOpenHelper(context);
     }
 
     private List<Recipe> getDailyRecipes() {
